move try_from_bits and FromBitsError to fallible module
