+++ enumflags/src/fallible.rs
@@ -1,5 +1,5 @@
 use core::convert::TryFrom;
-use super::{BitFlags, FromBitsError};
+use super::BitFlags;
 use super::_internal::RawBitFlags;
 
 macro_rules! impl_try_from {
@@ -16,7 +16,15 @@ macro_rules! impl_try_from {
             type Error = FromBitsError<T>;
 
             fn try_from(bits: T::Type) -> Result<Self, Self::Error> {
-                Self::try_from_bits(bits)
+                let flags = Self::from_bits_truncate(bits);
+                if flags.bits() == bits {
+                    Ok(flags)
+                } else {
+                    Err(FromBitsError {
+                        flags,
+                        invalid: bits & !flags.bits(),
+                    })
+                }
             }
         }
     };
@@ -24,4 +32,27 @@ macro_rules! impl_try_from {
 
 impl_try_from! {
     u8, u16, u32, u64, usize
+}
+
+#[derive(Debug, Copy, Clone)]
+pub struct FromBitsError<T: RawBitFlags> {
+    flags: BitFlags<T>,
+    invalid: T::Type,
+}
+
+impl<T: RawBitFlags> FromBitsError<T> {
+    pub fn truncate(self) -> BitFlags<T> {
+        self.flags
+    }
+
+    pub fn invalid_bits(self) -> T::Type {
+        self.invalid
+    }
+}
+
+#[cfg(feature = "std")]
+impl<T: RawBitFlags> std::error::Error for FromBitsError<T> {
+    fn description(&self) -> &str {
+        "invalid bit representation"
+    }
 }
